fix(oas, path): add _id suffix (#13)
The aep guidance is to add the _id suffix
on all path elements. For consistency, adding the _id
suffix to the pattern as well.

Making this change also required propagating the snake_case
convention to the resource singular and API conversion as well -
without doing so, conversion logic is necessary to extract and generate
the appropriate paths. 

This is aligned with the work to align with snake case
across the board: https://github.com/aep-dev/aeps/discussions/320

This change is being made in lockstep with
https://github.com/aep-dev/aeps/pull/321 to update the guidance.